+++ brayns/network/adapters/ArrayAdapter.h
@@ -95,7 +95,7 @@ struct ArrayAdapter
      */
     static bool deserialize(const JsonValue& json, T& value)
     {
-        auto array = JsonHelper::extractArray(json);
+        auto array = JsonExtractor::extractArray(json);
         if (!array)
         {
             return false;
