+++ brayns/network/json/JsonAdapter.h
@@ -36,13 +36,35 @@ namespace brayns
 template <typename T>
 struct JsonAdapter
 {
+    /**
+     * @brief Get the JSON schema using getSchema() method of value.
+     *
+     * @param value Value to get the schema from.
+     * @return JsonSchema JSON schema of value.
+     */
     static JsonSchema getSchema(const T& value) { return value.getSchema(); }
 
+    /**
+     * @brief Serialize value into json.
+     *
+     * @param value Input value.
+     * @param json Output JSON.
+     * @return true Success.
+     * @return false Failure.
+     */
     static bool serialize(const T& value, JsonValue& json)
     {
         return value.serialize(json);
     }
 
+    /**
+     * @brief Deserialize json into value.
+     *
+     * @param json Input JSON.
+     * @param value Output value.
+     * @return true Success.
+     * @return false Failure.
+     */
     static bool deserialize(const JsonValue& json, T& value)
     {
         return value.deserialize(json);
